cmd/server: remove Customer routes
